fix:Snake不能回头特性 - 特殊情况 Snake 长度为1时，可以回头 - getLength 改 int 为 size_t

+++ include/Snake.h
@@ -17,7 +17,7 @@ public:
   void decline();
   void move();
   void turn(Direction d);
-  int getLength() const { return segments.size(); };
+  size_t getLength() const { return segments.size(); };
   Point getHead() const { return segments[0]; };
   vector<Point> getSegments() const { return segments; };
 };
